Fix chat routing bug
Button in transits tab routing to chat in friends tab made the friends tab break under certain conditions. The chat file had to be moved to an independent folder.
